# app/api/files/[id]/download/route.ts
import { NextResponse } from 'next/server';
import { getCurrentUser } from '@/app/api/admin/_utils';
import { FileManager } from '@/lib/files/file-manager';

export const runtime = 'nodejs';

export async function GET(_request: Request, { params }: { params: { id: string } }) {
  try {
    const session = await getCurrentUser();
    if (!session?.tenantId) return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });

    const file = await FileManager.getReadableFileById(params.id, {
      tenantId: session.tenantId,
      uid: session.uid,
      role: session.role,
    });
    if (!file) return NextResponse.json({ error: 'File not found' }, { status: 404 });

    const downloadUrl = await FileManager.generateDownloadUrl(file.storagePath, session.tenantId);
    return NextResponse.json({ downloadUrl });
  } catch (error: any) {
    return NextResponse.json(
      { error: error?.message || 'Failed to generate download' },
      { status: 500 },
    );
  }
}