# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bizosto ERP API
  version: 1.0.0
  description: >-
    Multi-tenant API contract generated from App Router route handlers.
servers:
  - url: /
security:
  - BearerAuth: []
...
paths:
...
  /api/auth/send-otp:
    post:
      summary: POST /api/auth/send-otp
      tags:
        - Auth
      responses:
        '200':
          description: Request completed successfully.
        '401':
          description: Unauthorized or invalid session.
        '403':
          description: Forbidden for current tenant or role.
...
  /api/automation/workflows/{id}:
    put:
      summary: PUT /api/automation/workflows/{id}
      tags:
        - Core
      responses:
        '200':
          description: Request completed successfully.
        '401':
          description: Unauthorized or invalid session.
        '403':
          description: Forbidden for current tenant or role.
...
  /api/cron/generate-invoices:
    get:
      summary: GET /api/cron/generate-invoices
      tags:
        - Core
      responses:
        '200':
          description: Request completed successfully.
        '401':
          description: Unauthorized or invalid session.
        '403':
          description: Forbidden for current tenant or role.
...
  /api/files/{id}/download:
    get:
      summary: GET /api/files/{id}/download
      tags:
        - Core
      responses:
        '200':
          description: Request completed successfully.
        '401':
          description: Unauthorized or invalid session.
        '403':
          description: Forbidden for current tenant or role.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT